Moved PlayerOther And ChatName outside
